pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "miai"
version = "0.3.0"
description = "Local AI agent powered by Ollama, accessible via Slack and Web UI"
requires-python = ">=3.11"
dependencies = [
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.27.0",
    "aiohttp>=3.9.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.19.0",
    "chromadb>=0.4.22",
    "ddgs>=9.11.0",
    "unstructured>=0.12.0",
    "python-magic>=0.4.27",
    "bcrypt>=4.1.0",
    "cryptography>=42.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "pytest-aiohttp", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
